Adds a test stub, moves method impls to their own modules, minor fixes.
