+++ zebra-rpc/src/indexer/tests.rs
@@ -1,0 +1,1 @@
+mod vectors;
