+++ zebrad/tests/acceptance.rs
@@ -3178,14 +3178,14 @@ async fn trusted_chain_sync_handles_forks_correctly() -> Result<()> {
     use zebra_state::{ReadResponse, Response};
 
     let _init_guard = zebra_test::init();
-    let mut config = random_known_rpc_port_config(false, &Network::new_regtest(None))?;
+    let mut config = os_assigned_rpc_port_config(false, &Network::new_regtest(None))?;
     config.state.ephemeral = false;
     let network = config.network.network.clone();
-    let rpc_address = config.rpc.listen_addr.unwrap();
 
     let test_dir = testdir()?.with_config(&mut config)?;
 
     let mut child = test_dir.spawn_child(args!["start"])?;
+    let rpc_address = read_listen_addr_from_logs(&mut child, OPENED_RPC_ENDPOINT_MSG)?;
 
     tracing::info!("waiting for Zebra state cache to be opened");
